Improve NodeIteratorCoreNumberOfTrianglesMetric.
The vertex with minimum degree will be computed based on an array, that
is initialized with the degree of each vertex. In each iteration the
entry of the vertex with minimum degree will be set to 0, as it is
removed from the graph and therefore has no edges anymore. The entries
of the neighboring vertices will be decreased by 1.