Add more test graphs for triangle counting metrics
Firstly additional test graphs for triangle counting metrics were added.
This includes a simple hexagon network with 7 nodes, where the 7th nodes
is in the middle of the graph and is connected to all other nodes. The
second example graph which was added is simply half of the hexagon graph
described above. 

Furthermore moved the email-EU-core graph to another folder in order to
improve the test resource file structure. 

Lastly the AbstractNumberOfTrianglesMetricTest was adapted to reflect
the changes described above.