Improve code quality of ForwardNumberOfTriangleMetric
Greatly improved the code quality of the ForwardNumberOfTriangleMetric
by exporting all ordering related code to a custom class. This
simplifies the actual algorithm a lot. Furthermore moved the two classed
to a custom package in order to have a clearer separation of related
classes. All classes using the metric were adapted accordingly as well.