Adapt EvaluationRunner to simplify evaluation process
Adapted the EvaluationRunner such that the actual evaluation process
becomes easier for the user. Instead of waiting for the whole evaluation
to end before printing any results, the new process prints one file for
each graph which is evaluated. This offers the benefit that, incase an
evaluation fails at some graph, it can be restarted from that graph
without the need to rerun on all previous graphs. Furthermore the output
is a lot easier to interpret. 

The README was adapted accordingly.